Make explicit binding available when using other SLs
It seems that the Metal renderer expects a few buffer slots to work in a specific way in The Forge that to some degree will break with other shading languages like Slang.  FSL handles a lot of the binding generation for you here - but that's not ideal when you need runtime compiled shaders.  Make explicit binding available for these scenarios.  TBD on Vulkan and D3D12.